refactor(core): isolate document source evidence
